requirements.txt
flask
torch
torchvision
nltk
pillow
numpy==1.26.4
scipy
tqdm
tensorboard
protobuf